### 2021-07-23 #3528 @whedon
[PRE REVIEW]: elmada: Dynamic electricity carbon emission factors and prices for Europe
**Submitting author:** @mfleschutz (<a href="http://orcid.org/0000-0002-8516-9635">Markus Fleschutz</a>)
**Repository:** <a href="https://github.com/DrafProject/elmada" target ="_blank">https://github.com/DrafProject/elmada</a>
**Version:** v0.0.6
