### 2021-07-24 #3528 @danielskatz
I think this is still the best option for this submission.
