### 2021-08-10 #3528 @fraukewiese
@milicag , @ckaldemeyer – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is elmada: Dynamic electricity carbon emission factors and prices for Europe
