### 2021-08-12 b773fd0
Add bump2version and module meta data
