+++ .bumpversion.cfg
@@ -1,0 +1,8 @@
+[bumpversion]
+current_version = 0.0.6
+commit = True
+tag = True
+
+[bumpversion:file:setup.py]
+
+[bumpversion:file:elmada/__init__.py]
