### 2021-08-18 #3528 @fraukewiese
Great, thanks @olejandro !
We aim for the review process to be completed within about 4-6 weeks but it is necessary to make a start well ahead of this as JOSS reviews are by their nature iterative and any early feedback you may be able to provide to the author will be very helpful in meeting this schedule.
