### 2021-08-23 3edd48d
Update paper, openjournals/joss-reviews#3625
