### 2021-09-06 #3625 @mfleschutz
Thank you @nmstreethran for your valuable feedback in DrafProject/elmada/issues/1. I hope the testing options are clearer now.
