### 2021-09-22 #3625 @olejandro
@mfleschutz, a few comments/questions from me on the paper:

- In the statement of need you refer to `environmental assessment of power demands`. I would have used `environmental assessment of electricity supply` instead. Is there a particular reason for why you refer to demand rather than supply?
- In the summary you state `To properly assess this flexibility within energy system models, it is essential to know the wholesale price and associated carbon emissions per unit of electricity for a given time step.` Could you clarify why it is essential and the type of energy system models you refer to? For example, in energy system models like [Balmorel](https://github.com/balmorelcommunity/Balmorel) or [TIMES](https://github.com/etsap-TIMES/TIMES_model), electricity price and and associated carbon emissions are the results of modelling.
