### 2021-09-23 3b082b5
Update paper, openjournals/joss-reviews/#3625
