+++ paper/paper.bib
@@ -48,3 +48,16 @@
     urldate = {2021-07-10},
 	url = {https://www.watttime.org/aer/}
 }
+
+
+@article{Prina2020,
+  doi = {10.1016/j.rser.2020.109917},
+  url = {https://doi.org/10.1016/j.rser.2020.109917},
+  year = {2020},
+  publisher = {Elsevier {BV}},
+  volume = {129},
+  pages = {109917},
+  author = {Matteo Giacomo Prina and Giampaolo Manzolini and David Moser and Benedetto Nastasi and Wolfram Sparber},
+  title = {Classification and challenges of bottom-up energy system models - A review},
+  journal = {Renewable and Sustainable Energy Reviews}
+}
