### 2021-09-23 #3625 @mfleschutz
@olejandro, thank you for the valuable feedback.

> * In the statement of need you refer to `environmental assessment of power demands`. I would have used `environmental assessment of electricity supply` instead. Is there a particular reason for why you refer to demand rather than supply?
> * In the summary you state `To properly assess this flexibility within energy system models, it is essential to know the wholesale price and associated carbon emissions per unit of electricity for a given time step.` Could you clarify why it is essential and the type of energy system models you refer to? For example, in energy system models like [Balmorel](https://github.com/balmorelcommunity/Balmorel) or [TIMES](https://github.com/etsap-TIMES/TIMES_model), electricity price and and associated carbon emissions are the results of modelling.

I considered both points in the [latest changes of the paper](https://github.com/DrafProject/elmada/commit/3b082b53ed0e2fc2bbb8e9432799ddda17380eb7). Maybe some brief background: My modeling background lies in the optimization of the energy systems of individual industrial and commercial businesses. In such a low-level model, the intention is to evaluate the costs and emissions of electricity demand (see point 1) and the national electricity system is not modeled together with it, since only the prices and emission factors are of interest (see point 2).

> Another question @mfleschutz: for the functionality documentation, should one refer to the readme or is there another source?

The software functionality is documented in the readme. The methodology and the data are described in the [Applied Energy Paper](https://doi.org/10.1016/j.apenergy.2021.117040). Higher-level code is documented using doc-strings in the code. Type hints are used throughout the project.
An additional readthedocs-documentation was not seen as necessary due to the small number of high-level functions. It was omitted in favor of compact documentation so far. Of course, this might change as the project evolves. @olejandro, @nmstreethran, do you think it would add value to the project in the current state?
