### 2021-09-23 #3625 @nmstreethran
@mfleschutz I don't think a Read the Docs documentation is necessary at this point either. Since there are docstrings and hints available, the core functionality documented in the README is sufficient, in my opinion. It's definitely something to consider if you expand upon the functionality of the software in the future.

Apologies for the delay in completing the rest of my review. I'll get this done by tomorrow. Thanks!
