### 2021-09-29 #3625 @olejandro
Thank you @mfleschutz! The changes make it very clear now. Also the Applied Energy Paper (great work, btw!) is very useful in understanding the background.
One more comment I have on the **Statment of Need** - for the available functionality in elmada, how much is it complementary rather than an alternative to e.g. electricityMap?
Regarding the documentation, I find the current README sufficient. A few points that I've noticed:
- In the **Carbon Emissions factors** section, *European countries* is specified as a restriction. Are the included countries limited to AT, BE, CZ, DE, DK, ES, FI, FR, GB, GR, HU, IE, IT, LT, NL, PL, PT, RO, RS, SI or are other countries either in the EU or geographically in Europe included as well?
- `get_residual_load()` and `get_el_national_generation()` are not mentioned in the README.
