### 2021-09-29 #3625 @olejandro
@fraukewiese, need your guidance regarding `A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?`. The *target audience* is included in the summary, should @mfleschutz move/include it in the **Statement of Need** as well?
