### 2021-10-05 #3625 @fraukewiese
@olejandro @mfleschutz : I agree that the target audience should be mentioned in the Statement of need.
Furthermore, I agree to the comment from olejandro that some more reference to similar software/projects should be made in the paper explaining which similarities and which additional/different functionalities/features/aspects exist. ("One more comment I have on the Statment of Need - for the available functionality in elmada, how much is it complementary rather than an alternative to e.g. electricityMap?")
