### 2021-10-06 c925648
Add support for Europe30
With this commit, Elmada supports 30 European countries instead of the 20
originally analyzed in the APEN paper. In the APEN paper, countries were
excluded due to missing data, space restrictions, or less interesting analyses
e.g. for countries with only one conventional fuel type. Also, after the APEN
publication, more data became available. This commit makes little changes to
the code and documentation to include the 10 additional countries to the
elmada functionalities.

Changes:

- from_entsoe.prep_XEFs: use EU28 value if country not in Tranberg table
- Plotting: A
[…]
