+++ paper/paper.bib
@@ -24,13 +24,12 @@
     url = {https://github.com/DrafProject/elmada},
     urldate = {2021-07-23}
 }
-    
-    @misc{ElmadaZenodo,
+
+@misc{ElmadaZenodo,
     title = {Elmada},
     author = {Fleschutz, Markus},
     year = 2021,
-    doi = {10.5281/zenodo.5126771},
-    howpublished = {\url{https://doi.org/10.5281/zenodo.5126771}}
+    doi = {10.5281/zenodo.5126771}
 }
 
 @online{ElecMapApi,
