+++ doc/images/xef_country_map.py
@@ -1,0 +1,5 @@
+import elmada
+
+fig = elmada.plots.xef_country_map(year=2020, method="XEF_EP")
+fp = elmada.paths.BASE_DIR.parent / "doc/images/xef_country_map.svg"
+fig.write_image(str(fp))
