+++ elmada/paths.py
@@ -4,6 +4,7 @@ from typing import Optional
 from appdirs import user_cache_dir, user_config_dir
 
 from elmada.mode import is_safe_mode
+from elmada.mappings import COUNTRIES_FOR_ANALYSIS
 
 
 def _get_cache_directory() -> Path:
@@ -27,6 +28,7 @@ DATA_DIR = BASE_DIR / "data/raw"
 SAFE_CACHE_DIR = BASE_DIR / "data/safe_cache"
 
 
-def mode_dependent_cache_dir(year: Optional[int] = None):
+def mode_dependent_cache_dir(year: Optional[int] = None, country: Optional[str] = None):
     is_safe_year = year in range(2017, 2021) or year is None
-    return SAFE_CACHE_DIR if (is_safe_mode() and is_safe_year) else CACHE_DIR
+    is_safe_country = country in COUNTRIES_FOR_ANALYSIS
+    return SAFE_CACHE_DIR if (is_safe_mode() and is_safe_year and is_safe_country) else CACHE_DIR
