### 2021-10-06 74d60c3
Fix bug in mode_dependent_cache_dir
