### 2021-10-07 314d2f1
paper: expand statement of need
see openjournals/joss-reviews/issues/3625
