+++ paper/paper.bib
@@ -60,3 +60,28 @@
   title = {Classification and challenges of bottom-up energy system models - A review},
   journal = {Renewable and Sustainable Energy Reviews}
 }
+
+@article{Hawkes2010,
+ doi = {10.1016/j.enpol.2010.05.053},
+ author = {Hawkes, A. D.},
+ year = {2010},
+ title = {Estimating marginal {CO2} emissions rates for national electricity systems},
+ keywords = {CO2;electricity;marginal},
+ pages = {5977--5987},
+ volume = {38},
+ number = {10},
+ issn = {03014215},
+ journal = {Energy Policy}
+}
+
+@article{Tranberg2019,
+ author = {Tranberg, Bo and Corradi, Olivier and Lajoie, Bruno and Gibon, Thomas and Staffell, Iain and Andresen, Gorm Bruun},
+ year = {2019},
+ title = {{Real-time carbon accounting method for the European electricity markets}},
+ keywords = {Carbon accounting;Carbon emission;Carbon intensity;Flow tracing},
+ pages = {100367},
+ volume = {26},
+ issn = {2211467X},
+ journal = {{Energy Strategy Reviews}},
+ doi = {10.1016/j.esr.2019.100367},
+}
