### 2021-10-07 #3625 @mfleschutz
Thank you, @olejandro and @fraukewiese, for your fruitful comments. I respond by topics:

## Readme (`get_residual_load` and `get_el_national_generation`)

I added a table describing functions that provide pre-processed data, see [this commit](https://github.com/DrafProject/elmada/commit/c97610f60cde620707f7e3ef013ffaade18d949b).

## Geographic scope

Going through the question of geographic scope was especially insightful. With [this commit](https://github.com/DrafProject/elmada/commit/c925648739b842f1cf0c58a62bdfbd67b33cd4e5), `elmada` now supports 30 European countries instead of the 20 originally analyzed in the APEN paper. In the APEN paper, countries were excluded due to missing data, space restrictions, or less interesting analyses e.g. for countries with only one conventional fuel type. Also, after the APEN publication, more data became available. So I made changes to the code and documentation to include the 10 additional countries to the `elmada` functionalities.

## Paper

In [this commit](https://github.com/DrafProject/elmada/commit/314d2f14df450fa22a49a3903e27b6b201bacc7b) I expanded the Statement of Need section to elaborate, how `elmada` is a light no-cost alternative to commercial services with the focus on the modeling of decentral energy hubs. Also I removed the duplicate DOI.
