### 2021-10-11 #3625 @nmstreethran
Hi @mfleschutz @fraukewiese, I have completed my review and have a few minor comments, which you can find below. Overall, this is a very useful software package and I am happy to recommend to accept once these comments are addressed. I do apologise for the delay, and please let me know if any of my comments are unclear.

## environment.yml

- `entsoe-py` appears twice; the one under pip requirements can be removed.

## Paper

- In the summary, line 31, remove 'And' from the start of the sentence.
- The acronyms XEF and MEF are used in the Statement of Need (line 56); you should state what they stand for here instead of the section below.
- In line 67, state what the acronyms PP and PWL stand for.

## Documentation

- State elmada's target audience in the first paragraph.
- Specify that elmada uses [ISO 3166-1 alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2) in the README and add a link to its Wikipedia article so that users can look up country codes easily.
- Typos:
  - [36, 15] electricity
  - [64, 32] ENTSO-E
  - [68, 58] Involved
  - [88, 41] referred
  - [176, 9] API
