### 2021-10-12 1791978
Minors in documentation
see openjournals/joss-reviews/issues/3625
