### 2021-10-13 #3625 @mfleschutz
> @mfleschutz: In the paper, please check: Line 20: "elmada is developed in the open on GitHub". - "in the open" sounds strange to me, but I am not a native speaker, so just doublecheck, if this is a correct phrase.

Thanks for the comment. I saw this actually many times. Some examples: [https://developer.apple.com/swift](https://developer.apple.com/swift), [https://opensource.com/...](https://opensource.com/life/12/9/should-we-require-open-source-developed-openly), [this JOSS paper](https://joss.theoj.org/papers/10.21105/joss.00825).
