### 2021-10-13 99a71ca
Minor change in JOSS paper
see openjournals/joss-reviews/issues/3625
