### 2021-10-13 #3625 @mfleschutz
Thank you @fraukewiese, I just released `elmada v0.1.0` 🚀

The new zenodo DOI is [10.5281/zenodo.5566694](https://doi.org/10.5281/zenodo.5566694)

While checking the Zenodo metadata, I was wondering if Michael D. Murphy is to be included as an author. He contributed with [(non-code) active project direction](https://joss.readthedocs.io/en/latest/submitting.html#authorship) and is therefore co-author of the JOSS paper.
