### 2021-10-14 ca1659a
Include new Zenodo DOI in JOSS paper
see openjournals/joss-reviews/issues/3625
