+++ paper/paper.bib
@@ -29,7 +29,7 @@
     title = {Elmada},
     author = {Fleschutz, Markus},
     year = 2021,
-    doi = {10.5281/zenodo.5126771}
+    doi = {10.5281/zenodo.5566694}
 }
 
 @online{ElecMapApi,
