### 2021-10-18 9ab488e
Update readme after JOSS acceptance
