+++ README.md
@@ -29,16 +29,18 @@ The target group includes modelers of distributed energy hubs who need **el**ect
 
 ## Features
 
-* __Carbon emission factors (CEF)__ are calculated depending on country and year in up to quarter-hourly resolution.
-You can choose between
-  * grid mix emission factors (XEFs) from fuel type-specific [ENTSO-E] electricity generation data (`method="XEF_EP"`)
-  * and approximations using merit order based simulations which allow also for the calculation of marginal emission factors (MEFs).
-    The according Power Plant method (`PP`) and Piecewise Linear method (`PWL`) are described in [this open-access Applied Energy paper][APEN paper].
-    The data used depends on the method chosen, see [scheme below](#cef-scheme).
+* __Dynamic electricity Carbon Emission Factors (CEFs)__ are calculated depending on country and year in up to quarter-hourly resolution.
+There are two types of CEFs: __Grid Mix Emission Factors (XEFs)__ and __Marginal Emission Factors (MEFs)__.
+While XEFs reflect the carbon footprint of an electricity use (attributional approach), MEFs estimate the carbon impact (consequential approach) of a change in electricity demand (Learn more in the [white paper][CEFWhitepaper] from Tomorrow and WattTime).
+Choose between
+  * __XEFs__ from fuel type-specific [ENTSO-E] electricity generation data only for Germany (`XEF_EP`),
+  * and __XEFs__ & __MEFs__ from merit order based simulations for [30 European Countries][Europe30] (`XEF_PP`, `XEF_PWL`, `MEF_PP`, `MEF_PWL`).
+  The according Power Plant method (`PP`) and Piecewise Linear method (`PWL`) are described in the open-access [Applied Energy paper].
+  The data used depend on the method chosen, see [scheme below](#cef-scheme).
 
-* __Wholesale electricity prices__ are provided for European countries. You can choose between the real historical [ENTSO-E] data or the simulation results of the `PP` / `PWL` method.
+* __Wholesale electricity prices__ are provided for European countries. You can choose between the real historical [ENTSO-E] data (`hist_EP`) or the simulation results of the `PP` / `PWL` method.
 
-* Other interesting market data such as merit order lists, fuel-specific generation data, or power plant lists are provided as a by-product of the CEF calculations.
+* Other interesting market data such as merit order lists & plots, fuel-specific generation data, or power plant lists are provided as a by-product of the CEF calculations.
 
 ## Methodology
 
@@ -46,7 +48,7 @@ With the `XEF_EP` method, XEFs are calculated by multiplying the share matrix *S
 
 <img src="https://render.githubusercontent.com/render/math?math=\mathrm{XEF}^\mathrm{EP}_{t} = S_{t,f}\cdot\varepsilon_f">
 
-The methods `PP`, `PWL`, and `PWLv` are explained in [this Applied Energy paper][APEN paper]. Here is an overview:
+The methods `PP`, `PWL`, and `PWLv` are explained in the [Applied Energy paper]. Here is an overview:
  <!-- Converted from pptx via https://convertio.co/ -->
  <img src="https://github.com/DrafProject/elmada/raw/main/doc/images/scheme_CEF_calculation.svg" id='cef-scheme' width="900" alt="scheme_CEF_calculation">
 
@@ -63,16 +65,16 @@ The countries supported by `elmada` can be seen in the map below which is the ou
[…]
